+++ packages/core/src/defer/rendering.ts
@@ -7,7 +7,7 @@
  */
 
 import {CachedInjectorService} from '../cached_injector_service';
-import {NotificationSource} from '../change_detection/scheduling/zoneless_scheduling';
+import {NotificationSource} from '../change_detection/scheduling/change_detection_scheduling';
 import {EnvironmentInjector, InjectionToken, Injector, Provider} from '../di';
 import {
   DehydratedContainerView,
