+++ packages/core/src/pending_tasks.ts
@@ -11,7 +11,7 @@ import {ɵɵdefineInjectable} from './di/interface/defs';
 import {
   ChangeDetectionScheduler,
   NotificationSource,
-} from './change_detection/scheduling/zoneless_scheduling';
+} from './change_detection/scheduling/change_detection_scheduling';
 import {INTERNAL_APPLICATION_ERROR_HANDLER} from './error_handler';
 import {PendingTasksInternal} from './pending_tasks_internal';
