+++ packages/core/src/render3/component_ref.ts
@@ -12,7 +12,7 @@ import {ChangeDetectorRef} from '../change_detection/change_detector_ref';
 import {
   ChangeDetectionScheduler,
   NotificationSource,
-} from '../change_detection/scheduling/zoneless_scheduling';
+} from '../change_detection/scheduling/change_detection_scheduling';
 import {Injector} from '../di/injector';
 import {EnvironmentInjector} from '../di/r3_injector';
 import {RuntimeError, RuntimeErrorCode} from '../errors';
