+++ packages/core/src/render3/util/view_utils.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-import {NotificationSource} from '../../change_detection/scheduling/zoneless_scheduling';
+import {NotificationSource} from '../../change_detection/scheduling/change_detection_scheduling';
 import {RuntimeError, RuntimeErrorCode} from '../../errors';
 import {
   assertDefined,
