+++ packages/core/test/change_detection_scheduler_spec.ts
@@ -48,7 +48,7 @@ import {
 
 import {provideNoopAnimations} from '@angular/platform-browser/animations';
 import {isBrowser, isNode, withBody} from '@angular/private/testing';
-import {ChangeDetectionScheduler} from '../src/change_detection/scheduling/zoneless_scheduling';
+import {ChangeDetectionScheduler} from '../src/change_detection/scheduling/change_detection_scheduling';
 import {RuntimeError, RuntimeErrorCode} from '../src/errors';
 import {scheduleCallbackWithRafRace} from '../src/util/callback_scheduler';
 import {global} from '../src/util/global';
